requirements.txt
absl-py==2.1.0
alembic==1.13.1
astunparse==1.6.3
attrs==23.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.7.0
bokeh==3.5.0
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.4
contourpy==1.2.0
cryptography==3.4.8
cycler==0.12.1
dask==2024.6.2
dask-expr==1.1.6
distributed==2024.6.2
distro==1.7.0
emnist==0.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fasteners==0.14.1
Flask==3.0.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
flatbuffers==24.3.25
fonttools==4.49.0
fsspec==2024.6.1
future==0.18.2
gast==0.6.0
ghp-import==2.1.0
google-pasta==0.2.0
greenlet==3.0.2
grpcio==1.64.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.11.0
httplib2==0.20.2
idna==3.3
importlib_metadata==8.0.0
influxdb-client==1.39.0
integer==1.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
livereload==2.6.3
locket==1.0.0
lockfile==0.12.2
lxml==5.0.0
lz4==4.3.3
Mako==1.1.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.5.2
ml-dtypes==0.3.2
monotonic==1.6
more-itertools==8.10.0
msgpack==1.0.8
namex==0.0.8
netifaces==0.11.0
numpy==1.26.2
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
outcome==1.3.0.post0
packaging==23.1
pandas==2.1.4
paramiko==2.9.3
partd==1.4.2
pathspec==0.11.2
pexpect==4.8.0
Pillow==9.0.1
platformdirs==3.10.0
prettytable==2.5.0
protobuf==4.25.3
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
PuLP==2.8.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycparser==2.22
Pygments==2.18.0
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.1
PyYAML==6.0.1
pyyaml_env_tag==0.1
reactivex==4.0.4
requests==2.28.0
rich==13.7.1
scikit-learn==1.4.1.post1
scipy==1.12.0
SecretStorage==3.3.1
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.23
tblib==3.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.3.0
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.5
watchdog==3.0.0
wcwidth==0.2.5
webdriver-manager==4.0.1
Werkzeug==3.0.2
wrapt==1.16.0
wsproto==1.2.0
xyzservices==2024.6.0
zict==3.0.0
zipp==1.0.0
pytest>=8.0.0
